下单错误

+++ app/controllers/orders_controller.rb
@@ -46,7 +46,7 @@ class OrdersController < ApplicationController
       return redirect_to :back, notice: '业务不存在，请核对后重新下单' unless goods.present?
       is_on_sale?(goods.on_sale)
       price_current = current_user.my_price(goods.id).to_f
-      
+
       h_user = current_user.h_user
       # 批量订单
       if params[:multiple_order]
@@ -69,7 +69,7 @@ class OrdersController < ApplicationController
                 order.update_attributes(h_level_crrent: h_user.level_id, h_price_current: h_price_current)
               end
             end
-            multiple_order_total_price = total_count * price_current.to_i
+            multiple_order_total_price = total_count * price_current.to_f
             raise '余额不足，请充值后下单' if current_user.balance < multiple_order_total_price
             current_user.update_attribute(:balance, current_user.balance - multiple_order_total_price )
           end
@@ -117,7 +117,7 @@ class OrdersController < ApplicationController
       else
         remark = params[:order][:remark]
         price_current = goods.get_current_price(current_user.level_id)
-        total_price = price_current.to_i * count.to_i
+        total_price = price_current.to_f * count.to_i
 
         @order.update_attributes(goods_id: goods.id, price_current: price_current, count: count, total_price: total_price, remark: remark)
       end
